requirements.txt
psutil==5.4.7
Django==2.0.7
python-dotenv==0.10.1
psycopg2-binary==2.8.3
djangorestframework==3.10.2